% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{find_mle_bfgs}
\alias{find_mle_bfgs}
\title{Obtain the MLE via BFGS}
\usage{
find_mle_bfgs(design, outcome)
}
\arguments{
\item{design}{matrix, the design matrix containing covariates for the model.}

\item{outcome}{vector, the outcome variable for the model.}
}
\value{
A vector of linear regression coefficients.
}
\description{
This function finds the MLE estimate for the regression coefficients via BFGS.
}
