% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{High Performance Generalized Linear Model Constructor}
\usage{
hiper_glm(design, outcome)
}
\arguments{
\item{design}{matrix, the design matrix containing covariates for the model}

\item{outcome}{vector, the outcome variable for the model}
}
\value{
An object of class 'hglm'.
}
\description{
This function creates a high performance GLM model.
}
