% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{log_likelihood_gradient}
\alias{log_likelihood_gradient}
\title{Calculate the gradient of log likelihood under a linear model}
\usage{
log_likelihood_gradient(beta, design, outcome, noise_var = 1)
}
\arguments{
\item{beta}{vector, coefficients in the linear regression}

\item{design}{matrix, the design matrix containing covariates for the model.}

\item{outcome}{vector, the outcome variable for the model.}

\item{noise_var}{numeric, the variance of the random noise in the model. The default is 1.}
}
\value{
a vector of gradient of log likelihood
}
\description{
This function calculates the gradient vector of log likelihood for a linear regression
}
