% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_IO.R
\name{parse_options}
\alias{parse_options}
\title{Parse the Control Options}
\usage{
parse_options(option)
}
\arguments{
\item{option}{list, a list of control options passed to hiper_glm().}
}
\value{
An appropriate list of control options.
}
\description{
This function parses and validates the control options for hiper_glm().
The valid control options are described in hiper_glm().
Terminate the program if an option is invalid.
}
