% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_IO.R
\name{prepare_output}
\alias{prepare_output}
\title{Prepare the output}
\usage{
prepare_output(design, regression_coef)
}
\arguments{
\item{design}{matrix, the design matrix containing covariates for the model.}

\item{regression_coef}{vector, a vector of linear regression coefficients.}
}
\value{
An object of class hglm.
}
\description{
This function prepares the output for hiper_glm().
}
