% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{print.hglm}
\alias{print.hglm}
\title{Print Method for hglm}
\usage{
\method{print}{hglm}(hglm)
}
\arguments{
\item{hglm}{object, an object of class 'hglm'.}
}
\description{
Displays the hglm object.
}
