% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm_helper.R
\name{coef.hglm}
\alias{coef.hglm}
\title{Coefficients Extraction}
\usage{
\method{coef}{hglm}(hglm)
}
\arguments{
\item{hglm}{object, an object of class 'hglm'.}
}
\description{
This function extracts the coefficients from a hglm model.
}
