% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{vcov.hglm}
\alias{vcov.hglm}
\title{Variance-Covariance Matrix Calculation}
\usage{
\method{vcov}{hglm}(hglm)
}
\arguments{
\item{hglm}{object, an object of class 'hglm'.}
}
\description{
Variance-Covariance Matrix Calculation
}
